move favorites to users.controller
